Refactor "Job Timed Out" test
The online "Job Timed Out" test was testing the timeout behaviour of
our `wait` method and not actually the AWS Batch job timeout
(https://docs.aws.amazon.com/batch/latest/userguide/job_timeouts.html).
This meant it could be implemented with Mocking instead of an online
test.

The reason this refactoring occurred is that since we queried for log
events almost immediately after submitting the job. This resulted in a
chance that the log event stream could be created which would change
`events` from `nothing` to `LogEvent[]` and fail the test.