Wait for log events for certain tests
It appears that even when a job status becomes SUCCEEDED or FAILED the
output from the job may not have been written to the log stream yet.
As such we'll wait a duration for the logs to become available. Since we
don't precisely know what is going on here we'll also emit some warnings
to assist in further investigations.